Fix typo bug with vreinterpretq + CI improvements (#21)
* fix: Correct typo in use of ARM reinterpret u32 -> f32

The type order is swapped; we got it right most places but missed it
here. Not sure why CI didn't catch this error?

* chore: Add Ubuntu 24 to CI

* chore: Simplify matrix for build CI

* chore: Update nanobind pin

* chore: Hopefully fix CI with latest Python by using a proper env

* fix: Specify modern Python and remove unnecessary pip install

Eventually, we're going to want to add Python tests, so we might as well
set up for it. Also, nanobind only requires `typing_extensions` on
Python <3.11, so we can remove that install (which errors on Ubuntu 24
with Python 3.12).

* I love devops so much

https://github.com/actions/setup-python/issues/436#issuecomment-2200427387